+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [0.1.4]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [0.1.5]
+
+### Added
+
+- configurable test databases
+
+### Modified
+
+- test image upload without matplotlib
+
+## [0.1.4]
 
 ### Added